+++ deepr/utils/field.py
@@ -1,7 +1,7 @@
-"""Field"""
+"""Field."""
 
 from collections import namedtuple
-from typing import Any, Tuple
+from typing import Any, Tuple, Union
 
 import numpy as np
 import tensorflow as tf
@@ -17,16 +17,31 @@ class Field:
     dtype : tf.DType
         Tensorflow type of the field (automatically inferred if string)
     name : str
-        Name of the fields
+        Name of the field
+    sequence : bool
+        If True, the field represents a sequence.
+
+        Used for ``tf.Example`` message serialization : if ``sequence``
+        is ``True``, the field with be stored in the ``feature_list``
+        entry of a ``tf.train.SequenceExample``.
+
+        Automatically set if not given : ``True`` if ``shape``'s first
+        dimension is ``None``.
+
     shape : Tuple
         Shape of the field
     """
 
-    def __init__(self, name: str, shape: Tuple, dtype, default: Any = None):
+    def __init__(self, name: str, shape: Tuple, dtype, default: Any = None, sequence: bool = None):
         self.name = name
-        self.shape = shape
+        self.shape = tuple(shape)
         self.dtype = TensorType(dtype).tf
         self.default = default if default is not None else TensorType(dtype).default
+        self.sequence = sequence if sequence is not None else (self.shape[0] is None if shape else False)
+
+        if self.sequence and not self.shape:
+            msg = f"sequence=True but shape={self.shape}: expected at least one dimension."
+            raise ValueError(msg)
 
     def __repr__(self):
         return f"{self.__class__.__name__}({self.name}, {self.shape}, {self.dtype}, {self.default})"
@@ -34,24 +49,91 @@ class Field:
     def __str__(self):
         return self.name
 
+    @property
+    def feature_specs(self):
+        """Return feature specs for parsing Example messages."""
+        if self.sequence:
+            if any(dim is None for dim in self.shape[1:]):
+                return tf.io.VarLenFeature(dtype=self.dtype)
+            else:
+                return tf.io.FixedLenSequenceFeature(shape=self.shape[1:], dtype=self.dtype)
+        else:
+            if any(dim is None for dim in self.shape):
+                return tf.io.VarLenFeature(dtype=self.dtype)
+            else:
+                return tf.io.FixedLenFeature(shape=self.shape, dtype=self.dtype)
+
     def startswith(self, prefix: str):
         return self.name.startswith(prefix)
-
-    def has_fixed_len(self) -> bool:
-        return all(dim is not None for dim in self.shape)
-
-    def has_var_len(self) -> bool:
-        return any(dim is None for dim in self.shape)
 
     def as_placeholder(self, batch: bool = False) -> tf.placeholder:
         shape = tuple([None] + list(self.shape)) if batch else self.shape
         return tf.placeholder(dtype=self.dtype, shape=shape, name=self.name)
 
-    def as_feature(self):
-        if self.has_fixed_len():
-            return tf.FixedLenFeature(shape=self.shape, dtype=self.dtype, default_value=self.default)
+    def to_feature(self, value: np.array) -> Union[tf.train.Feature, tf.train.FeatureList]:
+        """Convert value to tf.train.Feature or tf.train.FeatureList.
+
+        For shapes with more than 2 dimensions, uses ``np.ravel`` to
+        flatten tensors in a list of values. Note that because
+        ``tf.Example`` uses row-major to parse list of values, we make
+        sure to use the same order with NumPy.
+
+        For that reason, if any of the dimensions is not set (i.e. is
+        ``None``), a ``ValueError`` is raised.
+
+        Parameters
+        ----------
+        value : np.array
+            Tensor values
+
+        Returns
+        -------
+        tf.train.FeatureList
+            If ``sequence`` is ``True``
+        tf.train.Feature
+            If ``sequence`` is ``False``
+
+        Raises
+        ------
+        ValueError
+            If ``sequence``, ``len(shape) > 2`` and one of the non-first
+            dimensions is not set (i.e. is ``None``).
+            If not ``sequence``, ``len(shape) > 2`` and any of the
+            dimensions is not set (i.e. is ``None``).
+        """
+
+        def _to_feature(val):
+            """Return tf.train.Feature"""
+            if self.dtype is tf.int32 or self.dtype is tf.int64:
+                return tf.train.Feature(int64_list=tf.train.Int64List(value=val))
+            if self.dtype is tf.float32 or self.dtype is tf.float64:
+                return tf.train.Feature(float_list=tf.train.FloatList(value=val))
+            if self.dtype is tf.string:
+                return tf.train.Feature(bytes_list=tf.train.BytesList(value=val))
+            else:
+                raise TypeError()
+
+        if self.sequence:
+            if len(self.shape) == 0:
+                msg = f"sequence=True but shape={self.shape}: expected at least one dimension."
+                raise ValueError(msg)
+            if len(self.shape) == 1:
+                return tf.train.FeatureList(feature=[_to_feature([val]) for val in value])
+            if len(self.shape) == 2:
+                return tf.train.FeatureList(feature=[_to_feature(val) for val in value])
+            if any(dim is None for dim in self.shape[1:]):
+                msg = f"Unable to convert field {self} to feature. If ndim > 2, dimensions must be static."
+                raise ValueError(msg)
+            return tf.train.FeatureList(feature=[_to_feature(np.ravel(val, order="C")) for val in value])
         else:
-            return tf.VarLenFeature(dtype=self.dtype)
+            if len(self.shape) == 0:
+                return _to_feature([value])
+            if len(self.shape) == 1:
+                return _to_feature(value)
+            if any(dim is None for dim in self.shape):
+                msg = f"Unable to convert field {self} to feature. If ndim > 2, dimensions must be static."
+                raise ValueError(msg)
+            return _to_feature(np.ravel(value, order="C"))
 
 
 _TensorType = namedtuple("TensorType", ("tf, np, py, default, string"))
@@ -70,6 +152,7 @@ _TENSOR_TYPES = [
 
 def TensorType(dtype):
     """Return TensorType from Python, TensorFlow or NumPy type"""
+    # pylint: disable=invalid-name
     for tt in _TENSOR_TYPES:
         if dtype is tt.tf:
             return tt